feat: ajout de la méthode updateInstrument

+++ src/main/java/fr/afpa/cda19/harmogestionapi/services/InstrumentService.java
@@ -1,7 +1,9 @@
 package fr.afpa.cda19.harmogestionapi.services;
 
 import fr.afpa.cda19.harmogestionapi.dtos.request.InstrumentCreateRequestDTO;
+import fr.afpa.cda19.harmogestionapi.dtos.request.InstrumentUpdateRequestDTO;
 import fr.afpa.cda19.harmogestionapi.dtos.response.InstrumentResponseDTO;
+import fr.afpa.cda19.harmogestionapi.exceptions.IdNonCorrespondantException;
 import fr.afpa.cda19.harmogestionapi.exceptions.RessourceDupliqueeException;
 import fr.afpa.cda19.harmogestionapi.exceptions.RessourceNonTrouveException;
 import fr.afpa.cda19.harmogestionapi.mappers.InstrumentMapper;
@@ -10,11 +12,10 @@ import fr.afpa.cda19.harmogestionapi.repositories.InstrumentRepository;
 import fr.afpa.cda19.harmogestionapi.repositories.PratiquerRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
-import java.util.Optional;
 
 /**
  * Service de liaison entre les contrôleurs et le repository
@@ -29,15 +30,18 @@ import java.util.Optional;
 public class InstrumentService {
 
     /**
-     * Repository des instruments
+     * Repository des instruments.
      */
     private final InstrumentRepository instrumentRepository;
 
     /**
-     * Repository de la pratique des instruments
+     * Repository de la pratique des instruments.
      */
     private final PratiquerRepository pratiquerRepository;
 
+    /**
+     * Mapper DTO<->entity des instruments.
+     */
     private final InstrumentMapper instrumentMapper;
 
     /**
@@ -45,6 +49,7 @@ public class InstrumentService {
      *
      * @param id l'identifiant de l'instrument recherché
      * @return ({@link InstrumentResponseDTO}) le dto correspondant à l'instrument recherché
+     * @throws RessourceNonTrouveException levée si l'instrument n'existe pas
      */
     public InstrumentResponseDTO getInstrument(int id) throws RessourceNonTrouveException {
         Instrument instrument = instrumentRepository.findById(id).orElseThrow(
@@ -61,6 +66,7 @@ public class InstrumentService {
         Iterable<Instrument> instruments = instrumentRepository.findAll();
         List<InstrumentResponseDTO> listeDto = new ArrayList<>();
         instruments.forEach(instrument -> listeDto.add(instrumentMapper.toDto(instrument)));
+        listeDto.sort(Comparator.comparingInt(InstrumentResponseDTO::id));
         return listeDto;
     }
 
@@ -68,28 +74,59 @@ public class InstrumentService {
      * Suppression d'un instrument par son identifiant.
      *
      * @param id l'identifiant de l'instrument à supprimer
+     * @throws RessourceNonTrouveException levée si l'instrument n'existe pas
      */
     @Transactional
-    public void deleteInstrument(int id) {
+    public void deleteInstrument(int id) throws RessourceNonTrouveException {
+        if (!instrumentRepository.existsById(id)) {
+            throw new RessourceNonTrouveException("Impossible de trouver l'instrument");
+        }
         pratiquerRepository.deleteAllByIdInstrument(id);
         instrumentRepository.deleteById(id);
     }
 
     /**
-     * Enregistrement ou modification d'un instrument.
+     * Enregistrement d'un nouvel instrument.
      *
-     * @param instrument l'instrument à créer ou modifier
-     * @return l'instrument après création ou modification
+     * @param dto ({@link InstrumentCreateRequestDTO}) le DTO contenant le nom de l'instrument à créer
+     * @return ({@link InstrumentResponseDTO}) le DTO correspondant à l'instrument créé
+     * @throws RessourceDupliqueeException levée en cas de duplication du nom de l'instrument
      */
-    @Transactional
-    public Instrument saveInstrument(Instrument instrument) {
-        return instrumentRepository.save(instrument);
-    }
-
     @Transactional
     public InstrumentResponseDTO createInstrument(InstrumentCreateRequestDTO dto)
             throws RessourceDupliqueeException {
-        if(instrumentRepository.existsByLibelleInstrument(dto.getNom())){
+        if (instrumentRepository.existsByLibelleInstrument(dto.getNom())) {
+            throw new RessourceDupliqueeException("Cet instrument existe déjà");
+        }
+        Instrument instrument = instrumentMapper.toEntity(dto);
+        Instrument savedInstrument = instrumentRepository.save(instrument);
+        return instrumentMapper.toDto(savedInstrument);
+    }
+
+    /**
+     * Mise à jour d'un instrument.
+     *
+     * @param dto ({@link InstrumentUpdateRequestDTO}) le DTO contenant le nom de l'instrument
+     *            à mettre à jour
+     * @param id  identifiant de l'instrument à mettre à jour en provenance de l'URL
+     * @return ({@link InstrumentResponseDTO}) le DTO correspondant à l'instrument mis à jour
+     * @throws RessourceNonTrouveException levée si l'instrument n'existe pas
+     * @throws IdNonCorrespondantException levée si l'id du DTO et le paramètre 'id' ne
+     *                                     correspondent pas
+     * @throws RessourceDupliqueeException levée en cas de duplication du nom de l'instrument
+     */
+    @Transactional
+    public InstrumentResponseDTO updateInstrument(final InstrumentUpdateRequestDTO dto,
+                                                  final Integer id)
+            throws RessourceNonTrouveException, IdNonCorrespondantException,
+            RessourceDupliqueeException {
+        if (!dto.getId().equals(id)) {
+            throw new IdNonCorrespondantException("Les identifiants ne correspondent pas");
+        }
+        if (!instrumentRepository.existsById(id)) {
+            throw new RessourceNonTrouveException("Impossible de trouver l'instrument");
+        }
+        if (instrumentRepository.existsByLibelleInstrumentAndIdInstrumentNot(dto.getNom(), id)) {
             throw new RessourceDupliqueeException("Cet instrument existe déjà");
         }
         Instrument instrument = instrumentMapper.toEntity(dto);
